Added pretty price formatter

+++ src/main/java/com/example/application/utils/fetchers/api_responses/AssetMetadata.java
@@ -150,7 +150,7 @@ public class AssetMetadata {
     private BigInteger lastBlockHashesPerSecond;
 
     @JsonProperty("LAST_BLOCK_DIFFICULTY")
-    private long lastBlockDifficulty;
+    private BigInteger lastBlockDifficulty;
 
     @JsonProperty("SUPPORTED_STANDARDS")
     private List<Object> supportedStandards;
